INT -> FLOAT

+++ camera/movements_layer.py
@@ -107,19 +107,19 @@ if __name__ == "__main__":
 
     elif sys.argv[1] == 'MOVE_HORIZONTAL':
         motor = 7
-        relative_degrees = int(sys.argv[2])
+        relative_degrees = float(sys.argv[2])
         move_motor(state, motor, relative_degrees)
 
     elif sys.argv[1] == 'MOVE_VERTICAL':
-        if int(sys.argv[2]) > 0:
+        if float(sys.argv[2]) > 0:
             # Means that the degrees are > 0
-            move_up(state, int(sys.argv[2]))
+            move_up(state, float(sys.argv[2]))
         else:
-            move_down(state, int(sys.argv[2]))
+            move_down(state, float(sys.argv[2]))
     
     elif sys.argv[1] == 'ORIENT_CLAW':
         motor = 6
-        relative_degrees = int(sys.argv[2])
+        relative_degrees = float(sys.argv[2])
         move_motor(state, motor, relative_degrees)
 
     elif sys.argv[1] == 'MOVE_CLAW':
